requirements.txt
Langchain
Langchain_community
Langchain_core
python-dotenv
streamlit
sse_starlette
bs4
pypdf
huggingface_hub
Langchain-huggingface
transformers


wikipedia 
arxiv
langchain_hub